migrating TFModelsSegmenter into tf-models module
