+++ eta/detectors/__init__.py
@@ -8,5 +8,5 @@ Brian Moore, brian@voxel51.com
 '''
 
 # Import all detectors into the `eta.detectors` namespace
-from .tfmodels_detectors import TFModelsDetector, TFModelsDetectorConfig
-from .tfmodels_segmenters import TFModelsSegmenter, TFModelsSegmenterConfig
+from .tfmodels_detectors import TFModelsDetector, TFModelsDetectorConfig, \
+                                TFModelsSegmenter, TFModelsSegmenterConfig
